# projects/expense-portal/src/openapi.yaml
openapi: "3.1.0"
info:
  title: Expense Management Portal
  description: Employee Expense Management Portal API — Acme Corporation
  version: "0.1.0"
  contact:
    name: Platform Engineering
servers:
  - url: https://expenses.acme.com
    description: Production (via Azure API Management)
paths:
...
  /v1/expenses/reports:
    get:
      operationId: listReports
      tags: [expenses]
      summary: List expense reports
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated list of reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListOut"
    post:
      operationId: createReport
      tags: [expenses]
      summary: Create a new expense report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreate"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummaryOut"
  /v1/expenses/reports/{report_id}:
    get:
      operationId: getReport
      tags: [expenses]
      summary: Get report details
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetailOut"
        "404":
          description: Not found
    patch:
      operationId: updateReport
      tags: [expenses]
      summary: Update a draft/rejected report
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated report
        "409":
          description: Report not editable
  /v1/expenses/reports/{report_id}/submit:
    post:
      operationId: submitReport
      tags: [expenses]
      summary: Submit report for approval
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRequest"
      responses:
        "200":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResponse"
        "422":
          description: Policy violations
  /v1/expenses/reports/{report_id}/line-items:
    post:
      operationId: addLineItem
      tags: [expenses]
      summary: Add a line item
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LineItemCreate"
      responses:
        "201":
          description: Line item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineItemOut"
  /v1/expenses/reports/{report_id}/line-items/{item_id}:
    patch:
      operationId: updateLineItem
      tags: [expenses]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LineItemUpdate"
      responses:
        "200":
          description: Updated line item
    delete:
      operationId: deleteLineItem
      tags: [expenses]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Line item deleted
...
  /v1/receipts/{receipt_id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Download receipt (SAS URL redirect)
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "302":
          description: Redirect to blob SAS URL
...
components:
  schemas:
...
    ReportCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        business_purpose:
          type: string
          maxLength: 1000
        submit:
          type: boolean
          default: false
      required: [title, start_date, end_date, business_purpose]
    ReportUpdate:
      type: object
      properties:
        title:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        business_purpose:
          type: string
    ReportSummaryOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        total_amount:
          type: number
        currency:
          type: string
        line_item_count:
          type: integer
        submitted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ReportListOut:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReportSummaryOut"
        next_cursor:
          type: string
          nullable: true
        total:
          type: integer
    ReportDetailOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/LineItemOut"
        approval_history:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalHistoryEntry"
    LineItemCreate:
      type: object
      properties:
        date:
          type: string
          format: date
        category:
          type: string
        vendor_name:
          type: string
          maxLength: 255
        amount:
          type: number
          exclusiveMinimum: 0
        currency:
          type: string
          pattern: "^(USD|CAD)$"
        description:
          type: string
          maxLength: 500
      required: [date, category, vendor_name, amount, currency, description]
    LineItemUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
        category:
          type: string
        vendor_name:
          type: string
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
    LineItemOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        category:
          type: string
        vendor_name:
          type: string
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        receipt_url:
          type: string
          nullable: true
        ocr_status:
          type: string
          nullable: true
...
    SubmitRequest:
      type: object
      properties:
        acknowledge_warnings:
          type: boolean
          default: false
    SubmitResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        submitted_at:
          type: string
          format: date-time
...
    ApprovalHistoryEntry:
      type: object
      properties:
        action:
          type: string
        actor:
          type: string
        timestamp:
          type: string
          format: date-time
        comment:
          type: string
          nullable: true
...